Wrap TemporalMemory core with "PairMemory" and "SequenceMemory"
This changes the Temporal Memory implementations to have the following
structure:

- A "TemporalMemory" class
  - The core of the implementation.
  - It doesn't have a compute method. It has no notion of timesteps --
    it has "activateCells" and "depolarizeCells" methods.
- A "PairMemory" subclass
  - Adds a compute method that calls "depolarizeCells" then
    "activateCells"
- A "SequenceMemory" subclass
  - Adds a compute method that calls "activateCells" and
    "depolarizeCells" methods.
  - Exposes a "getNextPredictedCells" method.
  - Automatically uses the activeCells and winnerCells as the basalInput.

You'll rarely use a TemporalMemory instance directly -- you'll either
instantiate a PairMemory or a SequenceMemory.

To use these via the network API, use the new ApicalTMPairRegion or
ApicalTMSequenceRegion. This change eliminates all other TM regions.

All-in all:

- This change makes it easier to use the TM correctly, with or without
  the Network API.
- This change makes it possible to get sequence memory predictions
  without incurring a big performance cost.
