+++ tests/algorithms/apical_tiebreak_temporal_memory_cpp/attm_cpp_apical_tiebreak_sequences_test.py
@@ -20,31 +20,30 @@
 # ----------------------------------------------------------------------
 
 """
-Run the apical tiebreak tests on the ExtendedTemporalMemory.
+Run the apical tiebreak sequence tests on the ExtendedTemporalMemory.
 """
 
 import unittest
 
-from htmresearch_core.experimental import ExtendedTemporalMemory
-from htmresearch.support.shared_tests.apical_tiebreak_test_base import (
-  ApicalTiebreakTestBase)
+from htmresearch_core.experimental import ApicalTiebreakSequenceMemory
+from htmresearch.support.shared_tests.apical_tiebreak_sequences_test_base import (
+  ApicalTiebreakSequencesTestBase)
 
 
-class ExtendedTM_ApicalTiebreakTests(ApicalTiebreakTestBase,
-                                     unittest.TestCase):
+
+class ExtendedTM_ApicalTiebreakSequencesTests(ApicalTiebreakSequencesTestBase,
+                                              unittest.TestCase):
   """
-  Run the apical tiebreak tests on the C++ ExtendedTemporalMemory.
+  Run the apical tiebreak sequence tests on the ExtendedTemporalMemory.
   """
 
-  def constructTM(self, columnCount, basalInputSize, apicalInputSize,
-                  cellsPerColumn, initialPermanence, connectedPermanence,
-                  minThreshold, sampleSize, permanenceIncrement,
-                  permanenceDecrement, predictedSegmentDecrement,
-                  activationThreshold, seed):
+  def constructTM(self, columnCount, apicalInputSize, cellsPerColumn,
+                  initialPermanence, connectedPermanence, minThreshold,
+                  sampleSize, permanenceIncrement, permanenceDecrement,
+                  predictedSegmentDecrement, activationThreshold, seed):
 
     params = {
       "columnCount": columnCount,
-      "basalInputSize": basalInputSize,
       "apicalInputSize": apicalInputSize,
       "cellsPerColumn": cellsPerColumn,
       "initialPermanence": initialPermanence,
@@ -53,22 +52,28 @@ class ExtendedTM_ApicalTiebreakTests(ApicalTiebreakTestBase,
       "sampleSize": sampleSize,
       "permanenceIncrement": permanenceIncrement,
       "permanenceDecrement": permanenceDecrement,
-      "predictedSegmentDecrement": predictedSegmentDecrement,
+      "basalPredictedSegmentDecrement": predictedSegmentDecrement,
       "activationThreshold": activationThreshold,
       "seed": seed,
       "learnOnOneCell": False,
     }
 
-    self.tm = ExtendedTemporalMemory(**params)
+    self.tm = ApicalTiebreakSequenceMemory(**params)
 
 
-  def compute(self, activeColumns, basalInput, apicalInput, learn):
+  def compute(self, activeColumns, apicalInput, learn):
 
     activeColumns = sorted(activeColumns)
-    basalInput = sorted(basalInput)
     apicalInput = sorted(apicalInput)
 
-    self.tm.compute(activeColumns, basalInput, apicalInput)
+    self.tm.compute(activeColumns,
+                    apicalInput=apicalInput,
+                    apicalGrowthCandidates=apicalInput,
+                    learn=learn)
+
+
+  def reset(self):
+    self.tm.reset()
 
 
   def getActiveCells(self):
